refactor(daemon): move mock/tooling crate inline tests to tests/ layouts
- wardnetd-mock: the synthetic_ip / synthetic_rtt tests from
  backends/noop_exit_probe.rs and backends/noop_latency_prober.rs merge
  into the existing src/tests/backends.rs suite.
- wardnet-test-agent: client/ping.rs regex tests move to a new
  src/client/tests/ping.rs (module-dir scaffold wired in client/mod.rs).
- wardnet-postupgrade: up/polkit.rs tests move to a new
  src/up/tests/polkit.rs (module-dir scaffold wired in up/mod.rs).

synthetic_ip / synthetic_rtt and the ping regexes are widened to
pub(crate); test bodies are unchanged.

Part of #847.
