+++ source/daemon/crates/wardnet-postupgrade/src/up/tests/mod.rs
@@ -1,0 +1,1 @@
+mod polkit;
