+++ source/daemon/crates/wardnet-test-agent/src/client/tests/mod.rs
@@ -1,0 +1,1 @@
+mod ping;
